test: migrate from test-workers
