+++ .github/workflows/ci.yml
@@ -243,29 +243,3 @@ jobs:
       - name: Run unit tests
         run: |
           npm run unit -w puppeteer-core -w puppeteer --if-present
-
-  workers-tests:
-    name: '[Required] Workers tests'
-    runs-on: ubuntu-latest
-    needs: check-changes
-    env:
-      CLOUDFLARE_API_TOKEN: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-      CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
-    steps:
-      - name: Check out repository
-        uses: actions/checkout@c85c95e3d7251135ab7dc9ce3241c5835cc595a9 # v3.5.3
-      - name: Set up Node.js
-        uses: actions/setup-node@e33196f7422957bea03ed53f6fbb155025ffc7b8 # v3.7.0
-        with:
-          cache: npm
-          node-version: lts/*
-      - name: Install dependencies
-        run: npm i
-        env:
-          PUPPETEER_SKIP_DOWNLOAD: true
-      - name: Run worker tests
-        run: |
-          cd ./test-workers
-          npm run build
-          npx wrangler deploy --compatibility-date $(date '+%Y-%m-%d')
-          npm run test
