--- test-workers/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "test-workers",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "deploy": "wrangler deploy",
-    "dev": "npx wrangler dev",
-    "start": "wrangler dev",
-    "test": "wireit",
-    "build": "wireit",
-    "cf-typegen": "wrangler types"
-  },
-  "wireit": {
-    "build": {
-      "command": "tsc",
-      "dependencies": [
-        "../packages/puppeteer:build"
-      ]
-    },
-    "test": {
-      "dependencies": [
-        "../packages/puppeteer:build"
-      ],
-      "command": "vitest"
-    }
-  },
-  "devDependencies": {
-    "typescript": "^5.5.4",
-    "wrangler": "^3.60.3"
-  }
-}
