--- test-workers/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "module": "esnext",
-    "target": "esnext",
-    "lib": ["esnext"],
-    "moduleResolution": "bundler",
-    "noEmit": true,
-    "skipLibCheck": true,
-    "allowSyntheticDefaultImports": true,
-    "rootDir": "src",
-    "types": ["./worker-configuration.d.ts"]
-  },
-  "include": ["./**/*.ts", "./**/*.d.ts"]
-}
