--- test-workers/worker-configuration.d.ts
@@ -1,11 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google Inc.
- * SPDX-License-Identifier: Apache-2.0
- */
-// Generated by Wrangler
-// by running `wrangler types`
-
-interface Env {
-  BROWSER: Fetcher;
-}
