--- test-workers/wrangler.toml
@@ -1,7 +1,0 @@
-
-name = "puppeteer-test-worker"
-main = "src/index.ts"
-compatibility_date = "2024-12-05"
-compatibility_flags = ["nodejs_compat"]
-
-browser = { binding = "BROWSER" }
